refactor(clippy): use clamp() and drop redundant casts (R-12)
Convert manual max/min clamp patterns to .clamp(), type the deep_verifier
score accumulators as f64 (dropping now-redundant casts). Compiles clean.

+++ core/crates/aisp-core/src/reference_validator/pipeline_verification.rs
@@ -66,8 +66,8 @@ impl<'a> PipelineVerifier<'a> {
         let mut proofs = Vec::new();
 
         for steps in test_cases {
-            let prose_rate = 0.62_f64.powi(steps as i32);
-            let aisp_rate = 0.98_f64.powi(steps as i32);
+            let prose_rate = 0.62_f64.powi(steps);
+            let aisp_rate = 0.98_f64.powi(steps);
             let improvement_factor = if prose_rate > 0.0 {
                 aisp_rate / prose_rate
             } else {
